vlib:  Swap src and dst for both move and swap ops
For move, instead of assigning the source value to the destination, swap
them in the shared code path.  Then update the *address* in the target's
parameter block (without actually changing any *values*) to point to the
source value object -- which now has the value originally in the target,
before it was swapped -- in order to call target_is_collectible(), using
the original destination value and type.