vx:  Link with dlmalloc (for MacRelix builds)
This allows use of Process Manager temporary memory to answer allocation
requests instead of drawing only from the application heap.  We're going
to need this as long as we depend on V's load() function for file input,
which loads the entire file at once, in order to support multi-megabyte-
sized files.