A-line:  Remove `rsrc` directive
This directive specifies a Mac resource file; however, it's not quite as
useful as it might seem.  In the process of porting A-line to target Mac
OS X, the resource copying machinery underlying `rsrc` was repurposed to
copy the result of `Rez`.  Since no projects were still using `rsrc`, it
was expedient to select a fork in which to store the resource data based
on the target platform.  Therefore, `rsrc` can't be used portably across
classic and bundled applications.  Classic development would necessitate
the use of MacCVS Pro (or another tool supporting MacBinary) in order to
store resource forks, and I'd rather not encourage workflows that depend
on defective version control systems.  It's possible to store a resource
file structure in the data fork for an application that's OS X only, but
since there's no native ResEdit on OS X, there's little reason to prefer
this over Rez source or separate files per resource.

In summary:  For now, use Rez.  In the future, there will be support for
resources as individual files.