Genie:  Add ".mbim" MacBinary+ mapping
Mac OS X's Finder is case-insensitive with regard to filename extensions
and doesn't distinguish between ".mbin" and ".mBin".  It's important for
MacBinary+ folder archives to be distinct from plain MacBinary archives,
since the MacBinary standard doesn't cover the former and most MacBinary
tools don't support it.

The final 'm' is mnemonic for "multiple", and indeed resembles a pair of
'n's.  Also, ".mbim" is more suggestive of ".mbin" than e.g. ".mbip" is.