Pedestal:  Draw the bundle name in the About box
Draw the application name (if compositing mode is available).  Leave the
window title empty in Carbon.  (This isn't so great for Carbon/CFM apps,
which won't have a single word of text in the window, but it's a work in
progess.)

Note:  Compositing mode is available in Mac OS X 10.2, and since earlier
versions of Mac OS X aren't supported, the compositing mode path applies
to Mac OS X generally (only counting Mach-O), as far as we're concerned.