MacBinaryDecoder:  Rewrite the owner resource
Mac OS 9's Finder will use this text as a fallback for the Version field
if 'vers' 1 is missing.  HIAboutBox will use it if CFBundleGetInfoString
is absent from Info.plist.  (OS X's Finder doesn't display this at all.)

It would be nice to append "\rAll rights reserved.", but the CR displays
as a box in Mac OS 9.  Anyway, it's a fallback in both cases, so it's no
big deal to override it.