Pedestal:  Set About menu item text dynamically
Shorten the stored About menu item text to "About...", but with a proper
ellipsis and a trailing NUL byte.  Use this as is if no application name
can be determined.  (The NUL byte won't be displayed.)

At run time, check for the closing NUL byte.  If absent, the application
has provided its own Apple menu resource, so leave it.  Otherwise, amend
the menu item text to include the application name.  Under Carbon, check
CFBundleName first.  Failing that, adopt the name of the owner resource.
If it doesn't exist or the name is empty, leave the existing text alone.