Genie:  Don't auto-create a bitmap in Draw()
While convenient for shell users, the timing dependency is a fountain of
confusion for shell scripters and programmers.  Instead, only create the
bitmap once the bounds/size has been set.  While this adds one more step
to an interactive scenario, interactive use is intended to be a stepping
stone to automation, not an end in itself, and removing the ambiguity of
when the bitmap is addressable helps users advance from experimenting to
building tools.