raster:  load/unload -> open/close + load, play
Rename load/unload to open/close.  Pass a flag argument to open_raster()
indicating whether the caller is requesting synchronized access:  If the
flag is clear, proceed as before; if it's set, then use a shared (rather
than private) memory mapping, so we can communicate with other processes
via concurrency management structures in shared memory, and disallow raw
Macintosh screen dump files (which don't have a metadata segment at all,
and therefore can't have the necessary synchronization relay) as well as
any rasters which require byte-swapping, since doing that (a) depends on
the privacy of the memory mapping so as not to clobber the metadata seen
by other processes, and (b) shouldn't be necessary, since by definition,
the processes sharing memory are on the same host.  Finally, reimplement
load_raster() as a wrapper around open_raster() without synchronization,
and add play_raster() as its complement.