Genie:  Don't auto-create gworld in Draw()
Create the GWorld when the bounds are set.  Creating a window is neither
necessary nor sufficient to create the GWorld anymore.

Clients desiring a specific bit depth should set that before setting the
bounds, to minimize reallocations.  If the depth isn't set explicitly in
advance, the system's default bit depth is used, which will typically be
the bit depth of the main screen.  (When multiple screens exist, it will
match at least one of them.)