Genie:  Add gworld stride property
This doesn't alter the GWorld itself, but it modifies all I/O operations
on its data.  It provides the client with a view of the pixel data as it
would be if the specified stride were applied instead of rowBytes.  This
is especially helpful with unpadded image data, since GWorlds always pad
rowBytes even when the actual row end is properly aligned in memory.  In
this iteration at least, the stride may not exceed rowBytes.