fbscreencap:  Handle 32-bit RGBA framebuffers
Previously, for 32 bits we assumed ARGB on big-endian platforms and BGRA
on little-endian ones, which was true enough for all variants of Mac OS,
as well as the Raspberry Pi and Nexus S.  However, the Nexus 4 uses RGBA
pixels.  In addition to detecting this and specifying Model_RGBx, create
the raster with big-endian metadata -- otherwise, a viewer program would
actually expect pixels in ABGR order and possibly byte-swap them back to
RGBA (if the display system doesn't handle ABGR natively, like QuickDraw
does), and the image would be misrendered in any case.