mac-qd-utils:  Add get_pix_rowBytes()
The 16-bit rowBytes field in struct PixMap has two flag bits, so only 14
bits are available for its stated purpose -- almost but not quite enough
for a row of 16384 bytes, which would be 4096 32-bit pixels.  A 27" iMac
with a high-DPI display already exceeds this.

Use GetPixRowBytes() (which lacks such a limitation) in Carbon, in which
it's always available.  Non-Carbon code, which even in Classic relies on
Mac OS 9, may continue to access rowBytes directly for now.  If it turns
out that Mac OS 9 can drive a 4K display, we'll reconsider.