Compile::Driver:  Honor `frameworks` directive
In the current semantics, a `frameworks` directive is only recognized in
the config of the executable in question -- it's not transitive like the
`use` directive.  If the set is empty, then Carbon is used as a default.
(These semantics may change in the future, but for now it's important to
be consistent with A-line.)