raster:  Make mb32 big-endian, move to a new unit
Let mb32('a', 'b', 'c', 'd') yield 'dcba' on little-endian platforms, so
it gets written to a file as "abcd".  Also, move the definition into its
own header so it's available for defining tags in the public interfaces.
('SCRN' never appears in a file, but make it consistent with future tags
which will.)