macos:  Check for rect src in sect_rect_region()
Sometimes the 3-arg sect_rect_region() is given a region which is a mere
rectangle, and simply ends after the ten-byte header.  Detect and handle
this by returning the input rectangle as the result region.  (It must be
equal to or enclosed by the source region, but this is guaranteed in our
use.)

Don't call the 5-arg sect_rect_region() in this case, since it expects a
proper region extent.  This fixes a bug where a rectangular intersection
of two non-rectangular regions comes out empty instead, resulting in the
failure to repaint the desktop after moving a window if the exposed area
is rectangular (which is possible if the window is partially offscreen).
Normally, this won't happen with document windows, which have a notch in
the upper-right and lower-left corners, ruining the rectangularity of an
exposed region when moving along an axis.  And dialog boxes, which don't
have notches, generally don't move.  But if a notch is clipped offscreen
and on the trailing edge when the window is moved (so that the remaining
notch is on the leading edge and doesn't leave an "eddy" in its "wake"),
then the exposed area will be rectangular.