vlib, vx:  Add secret() function
At the language level, secret() is an identity function that returns its
argument in all cases.  At the machine level, it causes the buffer owned
by its argument, if any, to be zeroed on deallocation -- provided that a
destructor isn't already set for the buffer.  For example, it works with
atomic values like strings and integers, but not aggregates like arrays.
However, it only works with strings and integers that are long enough to
require an allocated extent.  Strings of 32 bytes or more are guaranteed
to allocate an extent.