macos:  Set SaveUpdate/PaintWhite in PaintBehind()
However, set them *after* the first call to PaintOne().  This leaves the
choice to erase or not up to the caller.  In the case of _SizeWindow, we
definitely want to erase the exposed content area of the resized window.
For _MoveWindow (once we blit the visible window content into place), we
don't.  To handle the former case, we'll disagree with Inside Macintosh,
which reports, "PaintBehind clears the global variable PaintWhite before
calling PaintOne, so clobberedRgn isn't erased."  The caller can already
do this on its own if it wishes.

Inside Macintosh Volume I also says, "PaintWhite is reset after the call
to PaintOne."  However, it's not obvious whether it means the first call
to PaintOne() or the last.  The desktop isn't affected, but once we have
multiple windows, we'll have to consider this.  While it's convenient to
respect the caller's preference for the target window, the lower windows
should be treated uniformly and handled automatically -- if MoveWindow()
calls PaintBehind() with PaintWhite cleared, windows further back should
still have their exposed areas erased immediately.  Therefore, we re-set
PaintWhite at the earliest opportunity.

Also, set SaveUpdate to true after the first PaintOne() call.  A call to
SizeWindow() may specify false for the update flag if the application is
going to manage the update itself, but exposed windows behind the target
window probably need to be handled automatically.