macos:  Axe draw_one_byte()
Just call draw_masked_byte() with a mask of 0xFF.  Speed isn't a concern
here because draw_one_byte() is called at most three times in a row, and
only 1.5 times on average.  It just has to be correct.  Note that srcBic
in draw_one_byte() expects a pre-negated byte, but in draw_masked_byte()
it doesn't.