macos:  Fix race condition in WaitNextEvent()
If the loop body took long enough for `now` to exceed `future`, the wait
timeout got set to an extremely large value, and the application stalled
until a non-null event occurred.  Break the loop instead.

This hasn't been observed up to this point, but future implementation of
the event queue is sufficient to expose it.