macos:  Fix src addr miscalc in masked CopyBits()
We were effectively adding the delta of dstBits' and srcBits' respective
origins to the beginning of the blitting operation.

This had no effect on MoveWindow(), since the only display has an origin
of (0,0) and MoveWindow() uses the Window Manager port for blitting.  So
both srcBits and dstBits have an origin of (0,0).

However, when blitting from an offscreen buffer (with origin (0,0)) to a
window (whose top is always greater than the menu bar height), the delta
srcBits.bounds.top - dstBits.bounds.top is non-zero.  The bounds for the
window's bitmap are in port coordinates, so given an alias of screenBits
and an onscreen window, the portBits.bounds.top is negative.  The buffer
has a zero origin, so the delta is (0 - -y), where y is the window's top
in global coordinates.  This positive number is added to an offset which
is then *subtracted* from the index of the starting source row, with the
result that the source image begins and ends too early, and the image is
drawn too low, with the bottom clipped and garbage at the top.  The same
issue applies to dstBits.bounds.left, as well.

Windows whose content is entirely visible are copied by a different code
path and therefore weren't affected.