ams-core:  Fix _Gestalt trap
It turns out the _Gestalt trap itself actually returns two values -- the
error code in D0 and the actual result in A0.  Assigning the result into
the address passed as the 'response' argument is performed by glue code.

Unfortunately, we have no way specify that a function returns results in
multiple registers in C++.  So, we have to write the patch in assembler.