xv68k:  Add --ignore-screen-locks option
This is a replacement for ams-core's -Q switch, which will be removed so
that screen_lock doesn't depend on global state -- a prerequisite of the
future modularization of ams-core.

The --ignore-screen-locks and -Q options both control the same behavior,
but one is passed to xv68k and the other to ams-core; additionally, they
have inverse semantics.