vlib:  Add Value-derived class Term, for symbols
And, add further-derived classes Constant and Variable, for convenience.
Use these three classes instead of directly constructing Value instances
for symbols.  This removes the codependency between Value and Symbol, at
the cost of adding one between Symbol and Term.  (The latter pair can at
least be argued to be part of the same conceptual unit, and in any case,
the codependency can be resolved by moving clone() from Symbol to Term.)