vlib:  Distinguish nothing from NIL
Divide nothing into two states.  The first, NIL, retains the zero value.
The second keeps the name but gets a different value.

Nothing is the result of functions and operators (e.g. print and assert)
that return no value.  NIL is used internally to indicate the absence of
optional functionality, such as the types of unannotated terms (`var x`)
and the result of requesting a nonexistent operator overload handler, as
well as the results of type checking and coercion failures.