vlib:  Allow assignment to untargetable elements
An individual byte in a string can't be resolved to a pointer to a Value
object, because such objects don't exist -- rather, the bytes are packed
into a single object that can only be addressed as a whole.

Signal a failure to target an object by setting the `type` field to NULL
in the result.  This is only possible for the last step:  Given the code
`a[ b ][ c ] = d`, `a[ b ]` can be a packed structure, but `a` can't be.
(This is fine for now, because the only packed elements are bytes, which
don't contain subelements -- although we'll need to revisit this later.)

In the failure case, return the address of the packed container value in
the Target's `addr` field.  If the container provides the operator, call
the handler with both the Target and the original left operand (in which
the final targeting step is readily available).  If the operator remains
unhandled, throw an exception (as before).