render_xml.vobj:  Use entities, tweak a few things
Prefer entity references to character references (for the four big ones,
at least).  In general, rename things to gain similarity to the upcoming
render_html module.  Specifically, remove or factor out mentions of XML,
and rename escape() to escape_text(), since it's only used for text, not
attributes.  (In theory, attributes should also be escaped, but the only
one present in Info.plist is the `version` attribute on <plist> itself.)