render_xml, make-info-plist:  Refactor XML mapping
Replace the `[tag: attrs, content]` pattern with `tag: (attrs): content`
(or just `tag: content`, when no attributes are present).

The simpler code in make-info-plist.vx (as well as in render_at_depth())
speaks for itself.

This is confirmed to make no difference to MacBinaryDecoder's Info.plist
(which is the only significantly complex one).