vlib:  Protect block scope result on stack from GC
Exiting a block scope triggers a garbage collection pass, but the result
of invoking the block was vulnerable.  Temporarily link the result value
into the tracker's roots (i.e. the graph of reachable objects) while the
tracker culls unreachable objects.

(This didn't result in C++-level use-after-free, as reference counts are
still in effect, but it did cause closed-over symbols in closures to get
cleared.  While not an issue for many programs, it prevented coding in a
style that makes liberal use of function composition.)