[NEW] bignum:  Move bigint code previously in plus
This removes the dependency plus had on stack-chain (which was needed to
make recursive operations safe in classic Mac OS), which was conflicting
(though benignly) with stack-space, resulting in link warnings for `ln`.

(Specifically, _relix_stack_space() was duplicated, with the stack-space
implementation of it selected (as intended).  I'm not clear what exactly
would happen if stack-chain's had been used, though it possibly may have
just worked.  Anyway, most plus project clients have no need of checking
the stack space at all, so it's best not to burden them with it.)