Genie:  Check signals and loop in AESendBlocking()
Anticipate a thread waking before its Apple event reply is received.  It
should only happen in response to a signal, so check for that.

This fixes a crash in tlsrvr when the process is continued while waiting
for an Apple event reply (though this is masked by a failed assertion in
Process::Continue() in debug mode), when it calls AEGetParamPtr() on the
reply structure after assuming it had been populated with received data.