Vertice:  Axe incremental update modes
Get rid of the `gBlitting = false` render mode.  This gave us the option
of ray-tracing either entire faces or individual scanlines at a time, at
the cost of either slow or extremely slow rendering.  The purpose was to
provide ray-tracing ability while keeping the user informed as to what's
going on.  On modern hardware, though, rendering is relatively fast, but
actually pushing pixels onto the screen is quite slow, so an incremental
display just slows things down for no reason.  If incremental display is
desired for the purpose of debugging (or nostalgia), this is not the way
to do it.