Vertice:  Don't set the GWorld in DrawAnaglyphic()
That is, call paint_into_GWorld() instead of paint_into_thePort(), so we
don't have to save and restore the GWorld explicitly.  This is not quite
as efficient, since now there are additional restore and save operations
in between, but (a) this shouldn't make a noticeable difference, and (b)
the need to set the GWorld at all is going to be removed next.