mac-sys-utils:  Fix check for Process Manager
So, it turns out the poorly named gestaltLaunchControl bit indicates the
presence of the Process Manager.  To be fair, this is documented in both
Inside Macintosh, Volume VI and Inside Macintosh: Processes.  But within
Gestalt.h, its comment says "launch control support is available", which
is patent nonsense.  In fact, all of gestaltOSAttr's feature bits have a
comment that could have been derived solely from the identifier name.

This ends a mysterious and inscrutable MacRelix crash under MultiFinder,
in which _OSDispatch copes with an undefined selector by clobbering both
its own code and Apple's copyright message.