ams-core:  Split InitCursor() into two steps
To allow for a future unit split, we need to avoid touching any internal
state.  First, call set_cursor().  Either the cursor is visible and this
does everything we need, or it's hidden and we need to show it.  The new
init_cursor() routine tests for the latter case and forces the cursor to
be shown (even if it was multiply hidden), by zeroing CrsrState prior to
calling show_cursor().