mac-ui-utils:  Optimize deletion of all menu items
If the MenuRef isn't opaque, it's much simpler to empty the menu by just
zeroing three fields and resizing the handle (one trap call), than if we
make trap calls (that involve copying memory) in a loop.  As a bonus, if
we ever want to support the 64K ROM (which lacks DeleteMenuItem()), this
will actually work.