ams-core:  Run Button() loops efficiently
Deal with applications calling Button() (or StillDown()) in a tight loop
until the result changes.

The obvious implementation of Button() is to just call poll_user_input()
and return the button state.  (If you don't call poll_user_input(), then
the button state will never be updated, and you're in an infinite loop.)

However, this pegs the emulator's CPU usage at 100%.  To combat this, we
make repeated calls to Button() block until something happens.  This may
not be a change in button state -- any event, including a mouse movement
or key press, unblocks the call.  Only the first call in a loop is free,
and calling GetNextEvent() or WaitNextEvent() resets the in-a-loop flag.