[NEW] ams-pack:  Traps for packages 0 - 7
The PACK_N_patch() handler checks the appropriate slot in AppPacks for a
'PACK' code resource handle, and if NULL, tries to load one.  Therefore,
there are three ways to provide a package: (a) install the resource, (b)
populate the AppPacks slot before calling the package trap, or (c) patch
the package trap directly.