Genie:  Get fine-grained mouse movements on OS X
This hasn't been a problem on OS X 10.9 and 10.10, but it is on Mac OS X
10.4, 10.5, and 10.8, as well as 10.12.  Mac OS 9 also has this behavior
for Carbon applications.  Specifically, that even having set a one-pixel
mouse region for WaitNextEvent(), mouse-moved events are withheld for as
long as the mouse continues moving.

Install a mouse-moved event handler to keep an event tap view responsive
as the mouse is moved in Mach-O builds.  This change was ineffective for
Carbon applications in natively booted Mac OS 9, and Carbon applications
running natively in Mac OS X (as CFM) or in Classic don't have the issue
in the first place, so they're not included here.