xv68k:  Blacken the raster on exit
It's weird to see the AMS screen just freeze after an application quits.
Also, it's a privacy issue to leave a view of a user's activity (as well
as one's data, potentially) lying around.  Before terminating the raster
sync, paint it black.

If there isn't a raster sync, then either (a) we're running in MacRelix,
in which case the screen lives in a virtual file that's stored in memory
only, not on disk, or (b) this is presumably some kind of automated test
run in which the final graphics output must be retained for analysis, or
at least doesn't contain anything private.  (A third possibility is that
the screen isn't used at all and may not even exist in the first place.)