ams-core:  Fix ABA problem with Button() loops
When an application called Button() in a loop instead of WaitNextEvent()
or GetNextEvent(), it was possible for a click to be lost (and if it was
delivered as an atomic event, it definitely would be lost, since MBState
wouldn't be touched at all in that case).  Guarantee that a click yields
a true result from Button() at least once by using a flag to remember if
a mouse-down has happened since the last check.