v68k-callbacks:  Call update() directly
We were translating a memory address for the effect it had, as a form of
dependency injection -- since v68k-callbacks is a prerequisite of xv68k,
and so can't (mutually) depend on it.  But the screen update code is now
in v68k-screen, so just call it directly.  (This frees v68k-callbacks of
assumptions about screen size.)