vlib:  Add receiver type
A receiver is an object implementing the recv operator.  Objects created
by the `receiver` constructor implement recv by calling a function using
an empty argument list.  As the point is to receive data from somewhere,
it's perfectly acceptable for the function to block.  Use with caution a
function that always immediately returns a non-empty value.