vx:  Detect thread self-join and abort
This is a programming error.  Thread handle access implies a causal link
between the current thread and the target thread.  These links must only
ever be acyclic.  Ideally, a thread should never even see its own handle
(or that of any upstream thread), but at the very least we need to avoid
attempting to join the current thread.

This is an unrecoverable error.  Malicious code could use thread handles
to deadlock or crash the interpreter, so don't pass the `thread` routine
or any thread handles to untrusted code.