async-wakeup:  Loop in a deferred task in Mac OS X
Code which runs at interrupt-level in traditional Mac OS doesn't work in
quite the same way in Mac OS X.  One failure mode occurs in simultaneous
calls to PrimeTime(): one of them blocks waiting to acquire a mutex; the
other crashes.  Queueing a deferred task to call PrimeTime() rather than
calling it directly serializes the calls and is sufficient to ward off a
crash on PPC; unfortunately, the wakeup request still conflicts with the
timer (which tries to acquire the same mutex outside of PrimeTimeTask())
on x86, and a crash ensues.

However, we can put OS X's "interrupt-level" code semantics to good use.
Since a deferred task is a pthread in OS X, it's perfectly acceptable to
block for a while.  To request an async wakeup, install a deferred task.
In the deferred task body, call WakeUpProcess() in a loop until the flag
is cleared.  Call Delay() with an argument of 1 (about 16ms) for timing,
which busy-waits in classic Mac OS but sleeps in Mac OS X.

Remember:  Had WakeUpProcess() just been designed correctly in the first
place to terminate the current, /or if none, the **subsequent**/ call to
WaitNextEvent(), none of async-wakeup (whether Time Manager or pthreads)
would be necessary, and you wouldn't be reading this now.  Never forget.