Genie:  Fix serial port handle resource management
The NewSerialDeviceHandle() template was instantiating with Type matched
to value types rather than const references.  Aside from being uselessly
verbose in its resulting codegen, the SerialDeviceHandle instance's copy
construction was actively harmful, in that the subsequent destruction of
the copy of the other filehandle cleared the reference to that other, so
the next time we try to open this device, instead of sharing the other's
driver reference, we try to open the driver from scratch, which fails.

Example:

In one shell, run `cat /dev/tty.modem`.  It opens `.AOut` and `.AIn` and
then blocks in read().  In another shell, run `mcmd ati`.  This uses the
modem port's other device, `/dev/cu.modem`.  The existing driver pair is
used (during which the callup device is given priority over the other to
the actual underlying serial drivers).  However, a subsequent invocation
of `mcmd` would find that the reference to the other device's filehandle
was clear, and try to open the serial drivers itself (which would fail).
Eliminating the unwanted copy construction fixes the bug.