Pedestal:  Make default window-closed proc usable
If the window-closed proc hook is unset, call DestroyWindow() instead of
DisposeWindow() for non-Carbon builds.  This is consistent with what the
owned< WindowRef > disposer does, so it's not strictly necessary to keep
an owned< WindowRef > around for each window -- as long as the window is
visible onscreen and available for the user to close, the window list is
a sufficient ownership structure.