poseven:  Prevent double-join race condition
If two threads tried to join the same thread simultaneously, it's highly
likely that both would eventually call pthread_join().  The first has to
unlock the mutex before calling pthread_join() (which may block), giving
the other thread ample opportunity to lock the mutex and pass the status
checks.  Instead, add a new flag that protects against multiple joining.

One artifact of this patch is that a second (or subsequent) join attempt
will fail immediately, even while the target thread is still running, so
a secondary join can't be used to wait for a thread to complete.  Rather
than extend the semantics of join(), we'll add a new method for this use
case.  In the meantime, note that secondary joins previously resulted in
undefined behavior, so throwing early is not a regression.