poseven:  Add thread::wait()
While callers can now call join() multiple times without harm (resulting
in a thrown exception instead of undefined behavior), it's convenient to
be able to wait for a thread to complete without concern for whether any
other thread is doing the same.  thread::wait() will call pthread_join()
if the thread is joinable; otherwise, it waits on the condition variable
until the thread status indicates that it's terminated.