vlib:  Add fixed-length vector types
In addition to `vector[T](N)`, you can now write `vector[T, N]()`.  Both
create a vector of N elements of type T, with a zero-initialized buffer.
The difference is whether N is part of the type or not: vector[T](3) and
vector[T](4) are of the same type; vector[T, 3]() and vector[T, 4]() are
not.  A vector type which includes the length is a *fixed-length* vector
type.