relix, Genie:  Defer clearing ignorable signals
When setting the disposition of a signal to SIG_IGN, instead of clearing
a pending signal immediately, wait until delivery is attempted.  We need
to check on delivery in case the signal is pending on other threads than
the one calling sigaction().  Otherwise, a signal may persist as pending
on a thread that had blocked it prior to the call to sigaction() but had
subsequently unblocked it, so that changing the signal disposition would
finally result in delivery of the signal even though it should have been
discarded.  Since we check on delivery, checking in the sigaction() call
is redundant.

This is an observable change in behavior:  If a signal is blocked with a
signal action other than SIG_IGN, the signal is sent to the process (and
becomes pending), the signal action is set to SIG_IGN and then something
else, and finally the signal is unblocked, then the signal will actually
be delivered (whereas before it would have been discarded).  While SUSv2
mandates the old behavior, SUSv3 drops that requirement.  I imagine that
with the adoption of pthreads, it became inconvenient -- each thread has
its own set of pending signals, and a sigaction() call in one might have
to access the pending signal set in every thread -- an unwieldy prospect
for implementors.