Genie:  Don't actually stop OS threads on SIGSTOP
Leave a "stopped" thread to loop until a signal is received that must be
handled even if the process is stopped.  This includes SIGKILL (plus any
other fatal signals whose signal action is the default), and for reasons
that should be obvious, SIGCONT.  This change allows a stopped thread to
be either continued with SIGCONT, or killed with one of several signals,
including SIGHUP, SIGINT, SIGQUIT, SIGALRM, and SIGTERM.  Other possible
candidates were omitted because they apply to running code (e.g. SIGILL,
SIGABRT, and SIGPIPE), or their intended use case is to be caught rather
than terminate the process (SIGUSR1 and SIGUSR2).