Genie:  Separate pthread_kill() code path
Verify that the sending thread belongs to the same process as the target
thread.  Skip the PID == 1 check because PID 1 can only be the target if
it's the sender, and PID 1 never sends signals (or runs at all, for that
matter).  Don't check for stopped threads, because only AESendBlocking()
stops threads, and its callers don't create threads.