Genie:  Mark threads inactive after entry, breathe
Maintain the normal state of a thread as inactive (i.e. not found in the
set of active threads).  (Which threads are in the set doesn't matter as
such -- what's important is that when cooperatively scheduled work items
are pending, the set is non-empty, so the event loop knows to sleep only
briefly.)  Mark new threads active on creation and inactive on entry, to
ensure they get started.  Mark breathing threads active before the yield
point and inactive after.  No other occurrences are necessary.  Since we
keep threads normally inactive, we won't have to mark them as such while
stopping the underlying host threads, for example.