vx:  Test for thread cancellation in I/O, sleep
The rule is that we call thread::testcancel() anywhere a system call may
have been interrupted by our cancellation signal.  This includes all I/O
calls that check for errors (with print() being a notable exception that
we'll revisit another time), as well as sleep().  Note that the system()
Standard C library function retries waitpid() on EINTR errors; V's run()
and output-from-run() functions duplicate this behavior.  Therefore, you
can't interrupt a thread blocked on a coprocess (without a more advanced
facility than what exists).

Thread cancellation in its current form is intended for threads that are
blocked in system calls (or in a loop that makes such calls).  It has no
interruptive effect on threads waiting on a condition variable (in order
to send or receive on a channel, for instance) or executing freely.  For
the time being, V programs can close the channel in question (which will
cause channel I/O to fail or return immediately) in the former case, and
perhaps poll a flag in the latter.  Currently, attempting to cancel() an
uninterruptible thread will only block the cancel() call and consume CPU
until the target thread reaches a cancellation point or exits on its own
(or the entire process is terminated).