vlib:  Use less stack in 2-arg execute()
When calling a function that returns `Value`, store the result as `const
Value&`.  This is valid C++ and evidently eliminates a copy construction
(judging by the reduction in stack use and executable size).

This is necessary to keep the tests passing when running in xv68k, which
has a limited stack (and is highly sensitive to overruns).  Though we've
applied dynamic stack chaining to circumvent such faults while in deeply
recursive user-level functions, execute() is also recursive and consumes
a great deal of stack space (previously almost 800 bytes, now barely 700
bytes).  The next patch after this actually *reduces* the stack usage in
safe_calc(), and with that patch but not this one the wondrous number 27
test will crash, possibly because the needed stack extension is deferred
until too late (letting execute() overrun the stack) whereas had it been
done earlier, the newly attached stack space would suffice.  Ultimately,
V needs to dissociate its user-level stack from the C++ stack.  For now,
we wave a few chickens and sheepishly acknowledge our temporary success.