vlib:  Add min() and max() functions
These functions return the respective extremum of their arguments.  Note
that an array (or vector, or other sequence) is only one argument.  When
called with one argument (or zero), min() and max() return whatever list
was passed, even if the value passed doesn't support comparison.  To get
the actual minimum or maximum value in an array, dereference it and pass
the resulting list.