vlib:  Cull on scope exit for custom refs tracking
Although we can rely on a symbol's denoted type to limit the values that
it can store, not all base types are safe.  In vx, `thread` provides our
counterexample.  In such cases, check the value to see if its dispatched
methods include custom reference management support, and if so, consider
it collectible without checking the type.