vlib:  Don't cull on scope exit for trivial values
Rename symbol_list_is_collectible() to suggest that it's not a mere list
version of symbol_is_collectible().  The latter is called (from eval.cc)
to decide whether to track a symbol prior to assigning a value.  Whereas
the former is called (from execute.cc) to decide whether to perform a GC
run when exiting a block scope with local variables.  In fact, it's more
closely related to target_is_collectible(), which is called (again, from
eval.cc) prior to an assignment to determine whether to cull unreachable
objects afterward -- which already considers the currently stored value.

Modify the newly named symbol_list_with_values_is_collectible() to check
for trivial values in the same way.