cthread-custom:  Fix regression in MacRelix
The call to Resume() in thread_start() was removed, on the basis that it
did nothing not already covered by the thread's switch-in routine.  This
was in fact the case, and the switch-in routine was indeed called before
thread entry /if/ the Thread Manager was used.  However, this wasn't the
case with our custom threading implementation.

The solution is easy:  Call the switch-in routine on thread entry.  This
unbreaks a regression added in a83e5085c9c9e39f4e5820338c5476ccdcf91122.