cthread-either:  Let mac-config select a backend
Applications that either require System 7.5 or don't support features of
System 7 don't need to check for the Thread Manager.  They can assume it
exists or doesn't exist, respectively.

To override directly:  Define CONFIG_THREAD_MANAGER to 0 to suppress the
check and use custom threads.  Define CONFIG_THREAD_MANAGER_GRANTED to 1
to suppress the check and use the Thread Manager.