cthread-either:  [FIX] Assume PPC's system threads
In theory, the Thread Manager isn't guaranteed to be available, since we
could be running in System 7.1.2 without the Thread Manager add-on.  But
(a) in practice, System 7.1.2 is quite unlikely to be in use, and (b) we
don't have much choice, since custom threads are 68K only, so the reason
for checking is not to select an alternative but to fail gracefully.