vlib:  Extend map/ver/per blocks to use named args
An anonymous block given as the right operand to map, ver, or per is now
allowed to use certain single-letter parameter names rather than needing
to refer to the list of arguments `_`.  For map and ver, `{_^2}` can now
be written `{x^2}`, and the reducer `{head _ (+) tail _}` can instead be
written `{a (+) b}`.