vlib:  Return list assignment results, not input
For a list assignment like `const x, const y = 3, 4`, return the results
of the respective assignments as a new list, rather than the input list.
This matters if the operator is `:=` (approximation vs. duplication) and
the target list includes symbols with type annotations (which is not yet
possible at the moment -- a technicality that will shortly be remedied).